requirements.txt
annotated-types==0.7.0
blinker==1.9.0
cachetools==5.5.2
cairocffi==1.7.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
cors==1.0.1
dotenv==0.9.9
emoji==2.14.1
filelock==3.18.0
Flask==3.1.0
flask-cors==5.0.1
freetype-py==2.5.1
future==1.0.0
gevent==25.5.1
greenlet==3.2.2
grpcio==1.71.0
grpcio-status==1.71.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.0
MarkupSafe==3.0.2
nltk==3.9.1
numpy==2.2.5
pillow==11.2.1
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycairo==1.28.0
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pyparsing==3.2.3
PySocks==1.7.1
python-dotenv==1.1.0
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
rsa==4.9.1
textblob==0.19.0
tldextract==5.3.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
uharfbuzz==0.50.2
uritemplate==4.1.1
urllib3==2.4.0
Werkzeug==3.1.3
zope.event==5.0
zope.interface==7.2
